pyproject.toml
[project]
name = "ai-assistant"
version = "0.1.0"
description = "Local file and image summarisation with RAG capabilities"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.1.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "PyYAML>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["."]